Make firmware the default nginx server and remove IP addresses
When moving to the Vienna datacenter, the IP addresses of the new location were added to the firmware nginx configuration, so that nodes can load firmware updates without working DNS. Instead firmware is now made the default server, so any unmatched hostname/ip address will be treated as the firmware site. This way it works as expected regardless of the location and does not need to be adapted, if a IP address is added or changed.